requirements.txt
numpy==1.19
comet-ml==3.21.0
gpytorch==1.5.1
jupyter==1.0.0
kmeans-pytorch==0.3
matplotlib==3.3
mpi4py==3.1.1
notebook==6.2.0
PyYAML==5.4.1
scipy==1.7
torch==1.9.1
tqdm==4.62.3
pytorch-lightning==1.5.4
